Refactor references library
