+++ src/abis/safe/index.ts
@@ -1,0 +1,2 @@
+export * from './GnosisFactory';
+export * from './GnosisSafe';
