--- src/address/arb-one/vaults.ts
@@ -1,6 +1,0 @@
-enum ARBITRUM_MAINNET_VAULTS {
-  'USDC' = '0xbf11ac635fe97430b5333018778d0d746f67c989',
-  'ETH' = '0xd979e6d8a05ef11b185d2df3d76175eecae522af',
-}
-
-export default ARBITRUM_MAINNET_VAULTS;
