--- src/address/arb-sepolia/tokens.ts
@@ -1,6 +1,0 @@
-export enum ARBITRUM_SEPOLIA_TOKENS {
-  USDC = '0x3EF9d22C43ccE024405bc4a96580b9aE86e85121',
-  ETH = '0xC50A09E4ec9e61e1C143Eb24ee63fb4A1956E790'
-}
-
-export default ARBITRUM_SEPOLIA_TOKENS;
