--- src/address/arb-sepolia/vaults.ts
@@ -1,6 +1,0 @@
-enum ARBITRUM_SEPOLIA_VAULTS {
-  USDC = '0xfE438a1475816516e5E79cCAD5A31F34267107e5',
-  ETH = '0x709C85461127953507AAfDE3D21e3EFb07187ED0'
-}
-
-export default ARBITRUM_SEPOLIA_VAULTS;
