--- src/chain-configurations/chain-configurations.ts
@@ -1,10 +1,0 @@
-import { Chain } from '../chain';
-import IPythConfig from './interface/pyth-config.interface';
-import pythStableConfiguration from './pyth-stable-configuration';
-
-const chainConfigurations: Record<number, IPythConfig> = {
-  [Chain.ARBITRUM_SEPOLIA]: pythStableConfiguration,
-  [Chain.ARBITRUM_MAINNET]: pythStableConfiguration,
-};
-
-export default chainConfigurations;
