--- src/chain-configurations/enum/pyth-deployment.enum.ts
@@ -1,6 +1,0 @@
-enum PythDeployment {
-  STABLE = 'stable',
-  BETA = 'beta',
-}
-
-export default PythDeployment;
