--- src/chain-configurations/index.ts
@@ -1,6 +1,0 @@
-import pythStableConfiguration from './pyth-stable-configuration';
-import chainConfigurations from './chain-configurations';
-import PythDeployment from './enum/pyth-deployment.enum';
-import IPythConfig from './interface/pyth-config.interface';
-
-export { pythStableConfiguration, chainConfigurations, PythDeployment, IPythConfig };
