--- src/chain-configurations/interface/pyth-config.interface.ts
@@ -1,9 +1,0 @@
-import PythDeployment from "../enum/pyth-deployment.enum";
-
-interface IPythConfig {
-  priceIdBySymbol: Record<string, string>;
-  pythDeployment: PythDeployment;
-  publicEndpoint: string;
-}
-
-export default IPythConfig;
