--- src/collateral.ts
@@ -1,4 +1,0 @@
-export enum Collateral {
-  USDC = 'USDC',
-  ETH = 'ETH',
-}
