+++ src/constants/pyth.ts
@@ -1,0 +1,15 @@
+// PYTH
+// Values taken from https://pyth.network/developers/price-feed-ids#pyth-evm-stable
+export const PYTH_PRICE_ID_ETH = '0xff61491a931112ddf1bd8147cd1b641375f79f5825126d665480874634fd0ace';
+export const PYTH_PRICE_ID_BTC = '0xe62df6c8b4a85fe1a67db44dc12de5db330f7ac66b72dc658afedf0f4a415b43';
+export const PYTH_PRICE_ID_LINK = '0x8ac0c70fff57e9aefdf5edf44b51d62c2d433653cbb2cf5cc06bb115af04d221';
+export const PYTH_PRICE_ID_MATIC = '0x5de33a9112c2b700b8d30b8a3402c103578ccfa2765696471cc672bd5cf6ac52';
+export const PYTH_PRICE_ID_ARB = '0x3fa4252848f9f0a1480be62745a4629d9eb1322aebab8a791e344b3b9c1adcf5';
+export const PYTH_PRICE_ID_BNB = '0x2f95862b045670cd22bee3114c39763a4a08beeb663b145d283c31d7d1101c4f';
+export const PYTH_PRICE_ID_OP = '0x385f64d993f7b77d8182ed5003d97c60aa3361f3cecfe711544d2d59165e9bdf';
+export const PYTH_PRICE_ID_AVAX = '0x93da3352f9f1d105fdfe4971cfa80e9dd777bfc5d0f683ebb6e1294b92137bb7';
+export const PYTH_PRICE_ID_USDC = '0xeaa020c61cc479712813461ce153894a96a6c00b21ed0cfc2798d1f9a9e9c94a';
+export const PYTH_PRICE_ID_GMX = '0xb962539d0fcb272a494d65ea56f94851c2bcf8823935da05bd628916e2e9edbf';
+export const PYTH_PRICE_ID_SOL = '0xef0d8b6fda2ceba41da15d4095d1da392a0d2f8ed0c6c7bc0f4cfac8c280b56d';
+export const PYTH_PRICE_ID_XAU = '0x765d2ba906dbc32ca17cc11f5310a89e9ee1f6420508c63861f2f8ba4ee34bb2';
+export const PYTH_PRICE_ID_EUR = '0xa995d00bb36a63cef7fd2c287dc105fc8f3d93779f062f09551b0af3e81ec30b';
