+++ src/markets/index.ts
@@ -1,14 +1,15 @@
 import { encodeAbiParameters, keccak256, parseAbiParameters } from 'viem';
-import { Collateral } from './collateral';
-import { Chain } from './chain';
-import { markets, usedVaultsPerChain } from './constants';
+import { Chain, Collateral } from '../types';
+import { marketInfos } from '../constants/markets';
+import { usedVaultsPerChain } from '../constants/vaults';
 
-export const availableMarketsByEncodedPerChain: Record<
-  Chain,
-  Record<string, string>
-> = (Object.values(Chain).filter((c) => Number.isInteger(c)) as Chain[]).reduce(
+// Function used to provide Market metadata for the interface such as the Market Name, Market Symbol, 
+// and the type of market
+export const availableMarketsByEncodedPerChain: Record<Chain, Record<string, string>> = (
+  Object.values(Chain).filter((c) => Number.isInteger(c)) as Chain[]
+).reduce(
   (acc, chain) => {
-    acc[chain] = markets.reduce<Record<string, string>>((reverseMarkets, m) => {
+    acc[chain] = marketInfos.reduce<Record<string, string>>((reverseMarkets, m) => {
       Object.values(Collateral).forEach((c) => {
         if (usedVaultsPerChain[chain][c]) {
           const encoded = keccak256(
