--- src/orderStatus.ts
@@ -1,6 +1,0 @@
-export enum OrderStatus {
-  PENDING = 'PENDING',
-  CANCELLED = 'CANCELLED',
-  SETTLED = 'SETTLED',
-  EXPIRED = 'EXPIRED',
-}
