--- src/orderType.ts
@@ -1,7 +1,0 @@
-export enum OrderType {
-  OPEN_NEW_POSITION = 'OPEN_NEW_POSITION',
-  CLOSE_POSITION = 'CLOSE_POSITION',
-  INCREASE_POSITION = 'INCREASE_POSITION',
-  DECREASE_POSITION = 'DECREASE_POSITION',
-  CANCEL_ORDER = 'CANCEL_ORDER',
-}
