--- src/positionsStatus.ts
@@ -1,5 +1,0 @@
-export enum PositionStatus {
-  OPEN = 'OPEN',
-  CLOSED = 'CLOSED',
-  LIQUIDATED = 'LIQUIDATED',
-}
