+++ src/types/chain.ts
@@ -1,4 +1,3 @@
 export enum Chain {
-  ARBITRUM_SEPOLIA = 421614,
   ARBITRUM_MAINNET = 42161,
 }
